pyproject.toml
[tool.poetry]
name = "13th-homework"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
libgravatar = "^1.0.4"
bcrypt = "^4.1.3"
cloudinary = "^1.40.0"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
fastapi-limiter = "^0.1.6"
jinja2 = "^3.1.4"
redis = "^5.0.5"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--doctest-modules"
testpaths =[
    "tests", ]
pythonpath = "."